+++ locales/dateFnsContext.js
@@ -1,0 +1,3 @@
+import {createContext} from "react";
+const DateFnsLocaleContext = createContext()
+export default DateFnsLocaleContext