+++ src/MonthModeView.jsx
@@ -1,19 +1,21 @@
-import React, {useState} from 'react'
+import React, {useState, useContext} from 'react'
 import PropTypes from 'prop-types'
 import { format } from 'date-fns'
-import { useTheme, styled } from '@mui/material/styles'
+import {useTheme, styled, alpha} from '@mui/material/styles'
 import {
   Paper, Typography, Table, TableBody, TableCell, TableContainer,
   TableHead, TableRow, tableCellClasses
 } from "@mui/material"
+import { getDaysInMonth, isSameMonth } from 'date-fns'
 import EventNoteRoundedIcon from '@mui/icons-material/EventNoteRounded'
 import EventItem from "./EventItem.jsx"
+import {useTranslation} from "react-i18next"
 
 const StyledTableCell = styled(TableCell)(({ theme }) => ({
   [`&.${tableCellClasses.head}`]: {
-    borderTop: `1px solid #ccc !important`,
-    borderBottom: `1px solid #ccc !important`,
-    borderLeft: `1px solid #ccc !important`,
+    borderTop: `1px ${theme.palette.divider} solid !important`,
+    borderBottom: `1px ${theme.palette.divider} solid !important`,
+    borderLeft: `1px ${theme.palette.divider} solid !important`,
     ['&:nth-of-type(1)']: {
       borderLeft: `0px !important`
     }
@@ -25,7 +27,7 @@ const StyledTableCell = styled(TableCell)(({ theme }) => ({
     maxWidth: 64,
     cursor: 'pointer',
     verticalAlign: "top",
-    borderLeft: `1px solid #ccc`,
+    borderLeft: `1px ${theme.palette.divider} solid`,
     ['&:nth-of-type(7n+1)']: {
       borderLeft: 0
     },
@@ -46,9 +48,11 @@ const StyledTableRow = styled(TableRow)(({ theme }) => ({
 
 function MonthModeView (props) {
   const {
+    rows,
+    locale,
     options, 
     columns,
-    rows, 
+    legacyStyle,
     searchResult, 
     onTaskClick, 
     onCellClick, 
@@ -56,47 +60,37 @@ function MonthModeView (props) {
   } = props
   const theme = useTheme()
   const [state, setState] = useState({})
-  
-  /**
-   * @name onCellDragOver
-   * @param e
-   * @return void
-   */
+  const { t } = useTranslation(['common'])
+  const today = new Date()
+  let currentDaySx = {
+    display: 'block',
+    background: alpha(theme.palette.primary.main, 1),
+    borderRadius: '50%',
+    padding: '1px 3px',
+    color: '#fff',
+    width: 'fit-content',
+    margin: 'auto'
+  }
+
   const onCellDragOver = (e) => {
     e.preventDefault()
   }
-  
-  /**
-   * @name onCellDragStart
-   * @description
-   * @param e
-   * @param item
-   * @param rowIndex
-   * @return void
-   */
+
   const onCellDragStart = (e, item, rowIndex) => {
-    setState({...state, itemTransfert: {item, rowIndex}})
+    setState({
+      ...state,
+      itemTransfert: { item, rowIndex }
+    })
   }
-  
-  /**
-   * @name onCellDragEnter
-   * @description
-   * @param e
-   * @param elementId
-   * @param rowIndex
-   * @return void
-   */
+
   const onCellDragEnter = (e, elementId, rowIndex) => {
     e.preventDefault()
-    setState({...state, transfertTarget: {elementId, rowIndex}})
+    setState({
+      ...state,
+      transfertTarget: { elementId, rowIndex }
+    })
   }
-  
-  /**
-   * @name onCellDragEnd
-   * @description
-   * @param e
-   * @return void
-   */
+
   const onCellDragEnd = (e) => {
     e.preventDefault()
     if (!state.itemTransfert && !state.transfertTarget) return
@@ -106,50 +100,47 @@ function MonthModeView (props) {
     let rowInd = rowsCopy.findIndex(d => d.id === transfertTarget.rowIndex)
     
     if (rowInd !== -1) {
-      let dayInd = rowsCopy[rowInd]?.days?.findIndex(d => d.id === transfertTarget.elementId)
-      
+      let dayInd = rowsCopy[rowInd]
+        ?.days
+        ?.findIndex(d => d.id === transfertTarget.elementId)
       if (dayInd !== -1) {
         let day = rowsCopy[rowInd]?.days[dayInd]
         let splittedDate = transfert?.item?.date?.split('-')
-        
         if (!transfert?.item?.day) {
           // Get day of the date (DD)
           transfert.item.day = parseInt(splittedDate[2])
         }
-        
         if (transfert.item.day !== day?.day) {
           let itemCheck = day.data.findIndex(item => (
             item.day === transfert.item.day && item.label === transfert.item.label
           ))
-          
           if (itemCheck === -1) {
-            let prevDayEvents = rowsCopy[transfert.rowIndex].days.find(d => d.day === transfert.item.day)
-            let itemIndexToRemove = prevDayEvents?.data?.findIndex(i => i.id === transfert.item.id)
-            
+            let prevDayEvents = rowsCopy[transfert.rowIndex]
+              .days
+              .find(d => d.day === transfert.item.day)
+            let itemIndexToRemove = prevDayEvents
+              ?.data
+              ?.findIndex(i => i.id === transfert.item.id)
             if (itemIndexToRemove === undefined || itemIndexToRemove === -1) {
               return
             }
-            
             prevDayEvents?.data?.splice(itemIndexToRemove, 1)
             transfert.item.day = day?.day
             transfert.item.date = format(day?.date, 'yyyy-MM-dd')
             day.data.push(transfert.item)
-            setState({...state, rows: rowsCopy, itemTransfert: null, transfertTarget: null})
+            setState({
+              ...state,
+              rows: rowsCopy,
+              itemTransfert: null,
+              transfertTarget: null
+            })
             onEventsChange && onEventsChange(transfert.item)
           }
         }
       }
     }
   }
-  
-  /**
-   * @name handleCellClick
-   * @description
-   * @param event
-   * @param row
-   * @param day
-   * @return void
-   */
+
   const handleCellClick = (event, row, day) => {
     event.preventDefault()
     event.stopPropagation()
@@ -212,13 +203,13 @@ function MonthModeView (props) {
   }
   
   return (
-    <TableContainer component={Paper}>
+    <TableContainer component={Paper} sx={{ boxShadow: 'none' }}>
       <Table
         size="small"
         aria-label="simple table"
         stickyHeader sx={{ minWidth: options?.minWidth || 650 }}
       >
-        <TableHead sx={{height: 24}}>
+        {legacyStyle && <TableHead sx={{height: 24}}>
           <StyledTableRow>
             {columns?.map((column, index) => (
                 <StyledTableCell 
@@ -229,28 +220,49 @@ function MonthModeView (props) {
                 </StyledTableCell>
               ))}
           </StyledTableRow>
-        </TableHead>
+        </TableHead>}
         <TableBody>
           {rows?.map((row, index) => (
               <StyledTableRow
                 key={`row-${row.id}-${index}`}
-                sx={{ '&:last-child td, &:last-child th': { border: 0 } }}
+                sx={{ 
+                  '&:last-child th': { 
+                    border: 0,
+                    borderLeft: `1px ${theme.palette.divider} solid`,
+                    '&:firs-child': {
+                      borderLeft: 0
+                    }
+                  }
+                }}
               >
-                {row?.days?.map((day) => (
+                {row?.days?.map((day, indexD) => (
                   <StyledTableCell
                     scope="row"
                     align="center"
                     component="th"
-                    sx={{ px: 1 }}
+                    sx={{ px: 1, position: 'relative' }}
                     key={`day-${day.id}`}
                     onDragEnd={onCellDragEnd}
                     onDragOver={onCellDragOver}
                     onDragEnter={e => onCellDragEnter(e, day.id, row.id)}
                     onClick={(event) => handleCellClick(event, row, day)}
                   >
-                    <Typography variant="body2">{day.day}</Typography>
+                    {!legacyStyle &&
+                    index === 0 && columns[indexD]?.headerName?.toUpperCase()}.
+                    <Typography
+                      variant="body2"
+                      sx={day.day === getDaysInMonth(today) &&
+                      isSameMonth(day.date, today) &&
+                      currentDaySx || {}}
+                    >
+                      {day.day}
+                    </Typography>
                     {(day?.data?.length > 0 && renderTask(day?.data, row.id))}
-                    {/*<EventNoteRoundedIcon fontSize="large" htmlColor="#ccc" />*/}
+                    {legacyStyle && day?.data?.length === 0 &&
+                    <EventNoteRoundedIcon 
+                      fontSize="small" 
+                      htmlColor={theme.palette.divider} 
+                    />}
                   </StyledTableCell>
                 ))}
               </StyledTableRow>